+++ generic/GenericSearchAlgorithm.h
@@ -40,8 +40,8 @@ public:
 	virtual void GetPath(environment *env, state &from, state &to, std::vector<state> &path) = 0;
 	virtual void GetPath(environment *env, state &from, state &to, std::vector<action> &path) = 0;
 	virtual const char *GetName() = 0;
-	virtual int GetNodesExpanded() = 0;
-	virtual int GetNodesTouched() = 0;
+	virtual uint64_t GetNodesExpanded() = 0;
+	virtual uint64_t GetNodesTouched() = 0;
 	virtual void LogFinalStats(StatCollection *stats) = 0; 
 	virtual void OpenGLDraw() const {}
 };
